TreeList_Edit

+++ Foxoft/FormTreeView.cs
@@ -1,4 +1,5 @@
-﻿using DevExpress.XtraEditors;
+﻿using DevExpress.Utils.Menu;
+using DevExpress.XtraEditors;
 using DevExpress.XtraTreeList;
 using DevExpress.XtraTreeList.Menu;
 using DevExpress.XtraTreeList.Nodes;
@@ -38,8 +39,12 @@ namespace Foxoft
         {
             List<DcHierarchy> DcHierarchies = efMethods.SelectHierarchies();
             treeList1.DataSource = DcHierarchies;
+        }
 
-            //treeList1.Columns[treeList1.KeyFieldName].Visible = false;
+        private void treeList1_InitNewRow(object sender, TreeListInitNewRowEventArgs e)
+        {
+            string NewDocNum = efMethods.GetNextDocNum(false, "H", "HierarchyCode", "DcHierarchies", 4);
+            e.SetValue(treeListCol_HierarchyCode, NewDocNum);
         }
 
         private void treeList1_FocusedNodeChanged(object sender, FocusedNodeChangedEventArgs e)
@@ -50,23 +55,39 @@ namespace Foxoft
                 DcHierarchy = efMethods.SelectHierarchy(HierarchyCode.ToString());
         }
 
-        private void bbEdit_ItemClick(object sender, EventArgs e)
-        {
-            treeList1.OptionsBehavior.Editable = !treeList1.OptionsBehavior.Editable;
-            treeList1.ShowEditor();
-        }
-
         private void bbAddChild_ItemClick(object sender, EventArgs e)
         {
             TreeListNode newNode = treeList1.AppendNode(null, treeList1.FocusedNode);
-            newNode.SetValue(0, "New Node");
-            treeList1.OptionsBehavior.Editable = !treeList1.OptionsBehavior.Editable;
+            //newNode.SetValue("HierarchyCode", "");
+            treeList1.FocusedNode = newNode;
+
+            treeList1.OptionsBehavior.Editable = true;
+            treeList1.ShowEditor();
+        }
+
+        private void bbAdd_ItemClick(object sender, EventArgs e)
+        {
+            TreeListNode newNode = treeList1.AppendNode(null, treeList1.FocusedNode.ParentNode);
+            treeList1.FocusedNode = newNode;
+
+            treeList1.OptionsBehavior.Editable = true;
+            treeList1.ShowEditor();
+        }
+
+        private void bbEdit_ItemClick(object sender, EventArgs e)
+        {
+            treeList1.OptionsBehavior.Editable = true;
             treeList1.ShowEditor();
         }
 
         private void bbDelete_ItemClick(object sender, EventArgs e)
         {
-            treeList1.DeleteNode(treeList1.FocusedNode);
+            string hierarchyCode = treeList1.FocusedNode.GetValue(treeListCol_HierarchyCode)?.ToString();
+
+            int result = efMethods.DeleteHierarchy(hierarchyCode);
+
+            if (result > 0)
+                treeList1.DeleteNode(treeList1.FocusedNode);
         }
 
         private void treeList1_HiddenEditor(object sender, System.EventArgs e)
@@ -79,15 +100,50 @@ namespace Foxoft
             if (e.Menu is TreeListNodeMenu)
             {
                 treeList1.FocusedNode = ((TreeListNodeMenu)e.Menu).Node;
-                e.Menu.Items.Add(new DevExpress.Utils.Menu.DXMenuItem("Dəyiş", bbEdit_ItemClick));
-                e.Menu.Items.Add(new DevExpress.Utils.Menu.DXMenuItem("Əlavə Et", bbAddChild_ItemClick));
-                e.Menu.Items.Add(new DevExpress.Utils.Menu.DXMenuItem("Sil", bbDelete_ItemClick));
+
+                e.Menu.Items.Add(new DXMenuItem("Əlavə et", bbAdd_ItemClick, svgImageCollection1["add"], DXMenuItemPriority.Normal));
+                e.Menu.Items.Add(new DXMenuItem("Uşaq əlavə et", bbAddChild_ItemClick, svgImageCollection1["child"], DXMenuItemPriority.Normal));
+                e.Menu.Items.Add(new DXMenuItem("Dəyiş", bbEdit_ItemClick, svgImageCollection1["edit"], DXMenuItemPriority.Normal));
+                e.Menu.Items.Add(new DXMenuItem("Sil", bbDelete_ItemClick, svgImageCollection1["delete"], DXMenuItemPriority.Normal));
             }
         }
 
         private void treeList1_DoubleClick(object sender, EventArgs e)
         {
             DialogResult = DialogResult.OK;
+        }
+
+        private void treeList1_CellValueChanged(object sender, CellValueChangedEventArgs e)
+        {
+            string hierarchyCode = e.Node.GetValue(treeListCol_HierarchyCode)?.ToString();
+            string hierarchyDesc = e.Node.GetValue(treeListCol_HierarchyDesc)?.ToString();
+
+            if (!string.IsNullOrEmpty(hierarchyCode))
+            {
+                DcHierarchy = efMethods.SelectHierarchy(hierarchyCode);
+
+                if (DcHierarchy == null)
+                {
+                    DcHierarchy = new() { HierarchyCode = hierarchyCode, HierarchyDesc = "Yeni", HierarchyLevel = treeList1.FocusedNode.Level, HierarchyParentCode = treeList1.FocusedNode.ParentNode?.GetValue(treeListCol_HierarchyCode)?.ToString() };
+                    efMethods.InsertHierarchy(DcHierarchy);
+                }
+
+                if (!string.IsNullOrEmpty(hierarchyCode))
+                {
+                    efMethods.UpdateHierarchyDesc(hierarchyCode, hierarchyDesc);
+                }
+            }
+        }
+
+        private void FormTreeView_Activated(object sender, EventArgs e)
+        {
+            //AutoFocus FindPanel
+            if (treeList1 is not null)
+            {
+                treeList1.FindPanelVisible = false;
+                if (!treeList1.FindPanelVisible)
+                    treeList1.BeginInvoke(new Action(treeList1.ShowFindPanel));
+            }
         }
     }
 }
